validate kitchen access in delete, get, update side use cases.

+++ src/domain/use-cases/side/delete-side.ts
@@ -1,16 +1,23 @@
+import { User } from "../../entities";
+import { CustomError } from "../../errors";
 import { SideRepository } from "../../repositories";
 
 interface DeleteSideUseCase {
-  execute(sideId: number): Promise<object>;
+  execute(sideId: number, user: User): Promise<object>;
 }
 
 export class DeleteSide implements DeleteSideUseCase {
   constructor(private readonly sideRepository: SideRepository) {}
 
-  async execute(sideId: number): Promise<object> {
-    const side = await this.sideRepository.deleteBySide(sideId);
+  async execute(sideId: number, user: User): Promise<object> {
+    if (!user.rol) throw CustomError.unAurothorized('User role is required');
+    const sideFound = await this.sideRepository.getSideById(sideId);
+    if(sideFound.kitchenId !== user.kitchenId && user.rol !== 'SUPER_ADMIN') {
+      throw CustomError.unAurothorized('User does not have access to this kitchen');
+    }
+    const sideDeleted = await this.sideRepository.deleteBySide(sideId);
     return {
-      side: side
+      side: sideDeleted
     }
   }
 }
